refactor: removed circular dependencies
