+++ packages/network/src/provider/utils/rpc-mapper/methods/eth_signTransaction/eth_signTransaction.ts
@@ -1,5 +1,5 @@
 import type { ThorClient } from '../../../../../thor-client';
-import type { VeChainProvider } from '../../../../providers';
+import type { VeChainProvider } from '../../../../providers/vechain-provider';
 import {
     JSONRPCInternalError,
     JSONRPCInvalidParams,
