+++ packages/network/src/provider/utils/rpc-mapper/methods/eth_signTypedData_v4/eth_signTypedData_v4.ts
@@ -1,11 +1,11 @@
 import type { ThorClient } from '../../../../../thor-client';
-import type { VeChainProvider } from '../../../../providers';
+import type { VeChainProvider } from '../../../../providers/vechain-provider';
 import {
     JSONRPCInternalError,
     JSONRPCInvalidParams,
     stringifyData
 } from '@vechain/sdk-errors';
-import type { VeChainSigner } from '../../../../../signer';
+import type { VeChainSigner } from '../../../../../signer/signers';
 import { Address, type vechain_sdk_core_ethers } from '@vechain/sdk-core';
 
 /**
